Добавил таблицы. Поправил файлы конфигурации.

+++ db/scripts/012_ddl_alter_auto_post_add_file_id_table.sql
@@ -1,1 +1,1 @@
-ALTER TABLE auto_post ADD COLUMN file_id int REFERENCES files(id);
+ALTER TABLE auto_post ADD COLUMN file_id BIGINT REFERENCES files(id);
