+++ db/scripts/014_ddl_alter_auto_post_add_type_id_table.sql
@@ -1,1 +1,1 @@
-ALTER TABLE auto_post ADD COLUMN type_id int REFERENCES types(id);
+ALTER TABLE auto_post ADD COLUMN type_id BIGINT REFERENCES types(id);
